oprava index stranky a archivace pro dokumenty
