+++ webclient/dokument/templates/dokument/index_dokument.html
@@ -82,7 +82,6 @@
       {% trans "dokumenty.index.archivovat.text" %}
     </div>
   </div>
-</div>
 
   <script>
     // show / hide card description
